requirements.txt
mediapipe
numpy
openai
opencv_contrib_python
opencv_python
Pillow
protobuf
Requests
streamlit
